Initial layout refactoring
