+++ serverdensity/agent/checks.py
@@ -2528,7 +2528,7 @@ class checks:
 
                 if self.agentConfig.get('proxyUrl', '') != '':
                     self.mainLogger.info('doPostBack: using proxy: %s', self.agentConfig['proxyUrl'])
-                    
+
                     proxy = urllib2.ProxyHandler({'https': self.agentConfig['proxyUrl']})
                     opener = urllib2.build_opener(proxy)
                     urllib2.install_opener(opener)
